refactor(scripts): fondi create_dcl_analysis_issue + create_de_followup_issue in create_issue.py
Unifica i due script di creazione issue GitHub in un unico script
parametrizzato dalla variabile d'ambiente REPO.

- create_dcl_analysis_issue.py + create_de_followup_issue.py → create_issue.py
- Aggiornato validate-clean-catalog.yml per usare REPO env
- Test unificati e parametrizzati per entrambi i repo target

LOC nette: -60
